Rename interface header
